refactor: move react version test utilities to src/test/internal
