+++ src/test/index.ts
@@ -15,4 +15,3 @@ export {
 } from './react_warnings';
 export { sleep } from './sleep';
 export * from './emotion-prefix';
-export * from './react_version';
