+++ src/test/internal/index.ts
@@ -8,3 +8,4 @@
 
 export * from './render_custom_styles';
 export * from './test_custom_hook';
+export * from './react_version';
